Re #14739 Fix GCC errors

+++ MantidPlot/src/Mantid/MantidGroupPlotGenerator.cpp
@@ -186,7 +186,7 @@ bool MantidGroupPlotGenerator::groupIsAllMatrixWorkspaces(
 double MantidGroupPlotGenerator::getSingleLogValue(
     int wsIndex, const std::set<double> &logValues) const {
   double value = 0;
-  if (wsIndex < logValues.size()) {
+  if (wsIndex < static_cast<int>(logValues.size())) {
     auto it = logValues.begin();
     std::advance(it, wsIndex);
     value = *it;
